requirements.txt
openai>=1.26.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
pdfminer.six>=20221105
pillow>=10.0.0
pytesseract>=0.3.10
tiktoken>=0.5.0
supabase>=2.4.0
